Remove 'advanced mode'
The built-in structure cannot accomodate terraform projects of any
meaningful size, and most users who are able to edit the deployment
scripts would prefer to run from the CLI, so the 'advanced mode' has
lost it's purpose.

Instead, we'll add a config option to link to the source repo of the
deployment scripts.